+++ finetune/base.py
@@ -529,7 +529,6 @@ class BaseModel(object, metaclass=ABCMeta):
         os.environ['CUDA_VISIBLE_DEVICES'] = ",".join([str(gpu) for gpu in gpus])
         conf = tf.ConfigProto(allow_soft_placement=self.config.soft_device_placement,
                               log_device_placement=self.config.log_device_placement)
-        conf.gpu_options.allow_growth = True
         self.sess = tf.Session(config=conf)
 
     def _set_random_seed(self, seed=None):
